Move user avatar and timestamp into common module
